requirements.txt
alembic==1.4.2
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Deprecated==1.2.10
Flask==0.12.2
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
psycopg2==2.7.5
psycopg2-binary==2.8.4
PyGithub==1.51
PyJWT==1.7.1
python-dateutil==1.5
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
